Update 05-04_awards-history.md
The section on recipients of the Lynn A. Bonfield Award was removed following discussion with the Awards Committee Chair in 4/23 when it was decided that maintaining a list of recipients of the Lynn A. Bonfield Award recipients was not necessary. Note: I did not update the revision history because this change follows so closely to the previous change made in 3/23.